+++ lib/elastic/dsl/builders/filters/all.rb
@@ -1,4 +1,5 @@
 require 'elastic/dsl/builders/filters/geo'
+require 'elastic/dsl/builders/filters/range'
 
 module Elastic
   module DSL
@@ -7,6 +8,7 @@ module Elastic
 
         module All
           include Elastic::DSL::Builders::Filters::Geo
+          include Elastic::DSL::Builders::Filters::Range
         end
 
       end