+++ spec/elastic/dsl/spec_helper.rb
@@ -1,8 +1,0 @@
-require 'bundler/setup'
-Bundler.setup
-
-require 'elastic_dsl'
-
-RSpec.configure do |config|
-  # some (optional) config here
-end